@@ -40,8 +40,7 @@
   "lint-staged": {
     "*.{ts,js}": [
       "tslint --fix -c tslint.json",
-      "prettier --single-quote --parser typescript --print-width 120 --write './**/*.{ts,js}'",
-      "git add"
+      "prettier --single-quote --parser typescript --print-width 120 --write"
     ]
   },
   "keywords": [
@@ -55,34 +54,33 @@
   ],
   "authors": "Kabir Baidhya <kabirbaidhya@gmail.com>",
   "dependencies": {
-    "camelize": "^1.0.0",
+    "camelcase-keys": "^6.2.2",
     "debug": "4.1.1",
-    "knex": "0.20.11",
+    "knex": "0.20.14",
     "ramda": "^0.27.0",
-    "snakeize": "^0.1.0"
+    "snakecase-keys": "mesaugat/snakecase-keys#38efc7b"
   },
   "devDependencies": {
-    "@istanbuljs/nyc-config-typescript": "^0.1.3",
-    "@types/chai": "^4.2.3",
+    "@istanbuljs/nyc-config-typescript": "^1.0.1",
+    "@types/chai": "^4.2.11",
     "@types/chai-as-promised": "^7.1.2",
     "@types/debug": "^4.1.5",
     "@types/mocha": "^7.0.2",
+    "@types/node": "^13.11.1",
     "@types/ramda": "^0.27.3",
     "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
-    "husky": "^3.0.7",
-    "lint-staged": "^9.4.0",
+    "husky": "^4.2.5",
+    "lint-staged": "^10.1.3",
     "mocha": "^7.1.1",
-    "nyc": "^14.1.1",
-    "prettier": "^1.18.2",
-    "ts-node": "^8.4.1",
-    "tslint": "^5.20.0",
+    "nyc": "^15.0.1",
+    "prettier": "^2.0.4",
+    "source-map-support": "^0.5.16",
+    "ts-node": "^8.8.2",
+    "tslint": "^6.1.1",
     "tslint-config-leapfrog": "^1.0.3",
     "tslint-config-prettier": "^1.18.0",
     "tslint-no-unused-expression-chai": "^0.1.4",
-    "typescript": "^3.6.3"
-  },
-  "publishConfig": {
-    "access": "public"
+    "typescript": "^3.8.3"
   }
 }
